requirements.txt
altair==5.3.0
astor==0.8.1
attrs==23.2.0
backports.zoneinfo==0.2.1
base58==2.1.1
blinker==1.8.2
boto3==1.34.108
botocore==1.34.108
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cmake==3.29.3
colorama==0.4.6
Cython==3.0.10
enum-compat==0.0.3
et-xmlfile==1.1.0
exceptiongroup==1.2.1
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
idna==3.7
importlib-resources==6.4.0
jinja2==3.1.4
jmespath==1.0.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
numpy==1.24.4
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==24.0
pandas==2.0.3
# rem pillow==10.3.0
# 2406 pip-review==1.3.0
pkgutil-resolve-name==1.3.10
plotly-express==0.4.1
# 2406 pipuninstallpip
SQLAlchemy==2.0.31
protobuf==4.25.3
psutil==5.9.8
pyarrow==16.1.0
pycparser==2.22
pydeck==0.9.1
pygments==2.18.0
Pympler==1.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
referencing==0.35.1
requests==2.31.0
rich==13.7.1
rpds-py==0.18.1
s3transfer==0.10.1
selenium==4.20.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
streamlit==1.34.0
tenacity==8.3.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
trio==0.25.1
trio-websocket==0.11.1
typing-extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
urllib3==1.26.18
validators==0.28.1
watchdog==4.0.0
webdriver-manager==4.0.1
wsproto==1.2.0
zipp==3.18.2
cdifflib==1.2.6
streamlit_authenticator==0.3.2
pyyaml==6.0.0
